+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## 1.21.1
-### Feature
+## 1.22.0
+### Features
 * add sourcemaps to min js and css files [#430](https://github.com/jcubic/jquery.terminal/issues/430)
 * new option holdRepeatTimeout - which is number of the delay between keypress repeat when holding key
+* selection to change background color based on formatting like in Bash
 ### Bugfix
 * embed emoji regex by Mathias Bynens for better emoji detection
 * update typescript definition to new options
@@ -10,7 +11,6 @@
 * fix for combine characters
 * fix typescript definition for prompt and greetings [#425](https://github.com/jcubic/jquery.terminal/issues/425)
 * fix typo in holdTimeout option name
-
 
 ## 1.21.0
 ### Security